+++ example/config.js
@@ -2,4 +2,8 @@ const { createConfig } = require('larana-js')
 
 module.exports = createConfig({
 	defaultTheme: 'dark',
+	debug: false,
+	debugOptions: {
+		renderOutline: true,
+	},
 })