+++ src/shared/index.js
@@ -1,3 +1,4 @@
 module.exports = {
 	...require('./readonly-property.js'),
+	...require('./merge-deep.js'),
 }